pocket-pro-sba-selfhosted/backend/requirements_updated.txt
# Core Flask dependencies - Render.com optimized for Python 3.11+ (NO GEVENT)
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
flask-sqlalchemy==3.1.1
gunicorn==21.2.0

# AI and Vector Database
chromadb==0.4.24
google-generativeai>=0.4.1
langchain==0.1.14
langchain-google-genai==0.0.11
sentence-transformers==2.2.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
numpy==1.26.2

# Document processing
pypdf2==3.0.1
python-docx==1.1.0
markdown==3.5.1

# Additional dependencies for Render
pydantic==1.10.9
urllib3>=1.26.0,<3.0.0
setuptools>=68.0.0
wheel>=0.42.0
werkzeug==3.0.1
pytest==7.4.3 # Keep pytest for testing
